Unify partition creation behind strategy-based helpers in _partitions.
